+++ CHANGELOG.md
@@ -10,3 +10,8 @@
 
 - Diffrent note names added.
 - Piano buttons callback added.
+
+## 1.0.0
+
+- Black button width added
+- Example sound method changed